refactor: remove isolated `test` folder (#1027)
