+++ apps/calendar/src/slices/view.spec.ts
