+++ apps/calendar/src/utils/collection.spec.ts
