+++ apps/calendar/src/utils/cssHelper.spec.ts
