+++ apps/calendar/src/utils/dom.spec.ts
