+++ apps/calendar/src/utils/domEvent.spec.ts
