+++ apps/calendar/src/utils/math.spec.ts
