--- apps/calendar/test/app/month.spec.ts
@@ -1,10 +1,0 @@
-import Month from '@src/factory/month';
-
-describe('View/Month', () => {
-  it('Month view', () => {
-    // import test
-    const month = new Month(document.createElement('div'));
-
-    expect(month.render()).toBeInstanceOf(Month);
-  });
-});
